+++ assets/js/modules/getMoreProducts.mjs
@@ -1,0 +1,12 @@
+import getAllProducts from './getAllProducts.mjs';
+
+const getMoreProducts = (page) => {
+  const buttonLoadMore = document.querySelector('.button-moreproducts');
+  buttonLoadMore.onclick = (event) => {
+    event.preventDefault();
+    page += 1;
+    getAllProducts(page);
+  };
+};
+
+export default getMoreProducts;
