+++ assets/js/scripts.js
@@ -1,62 +1,9 @@
-const cardsGroup = document.querySelector('.cards-group');
-const buttonLoadMore = document.querySelector('.button-moreproducts');
-let currentPage = 1;
+import getAllProducts from './modules/getAllProducts.mjs';
+import getMoreProducts from './modules/getMoreProducts.mjs';
 
-buttonLoadMore.onclick = (event) => {
-  event.preventDefault();
-  currentPage += 1;
-  getProducts(currentPage);
-};
-
-const formatter = new Intl.NumberFormat('pt-BR', {
-  style: 'currency',
-  currency: 'BRL',
-});
-
-const getProducts = (page = 1) => {
-  const request = new XMLHttpRequest();
-  const urlAPI = `https://frontend-intern-challenge-api.iurykrieger.now.sh/products?page=${page}`;
-  const requestURL = urlAPI;
-  request.open('GET', requestURL);
-  request.responseType = 'json';
-  request.send();
-  request.onload = function () {
-    const arrayProducts = request.response;
-    getAllProducts(arrayProducts.products);
-  };
-};
+let page = 1;
 
 window.onload = () => {
-  getProducts();
-};
-
-const buildCard = (data) => {
-  return `
-    <div class="card">
-        <div class="card-header">
-            <img src="http:${data.image}">
-        </div>
-        <div class="card-content">
-            <h3 class="card-content__title">${data.name}</h3>
-            <p class="card-content__description">${data.description}</p>
-            <p class="card-content__prices__oldprice">De: ${formatter.format(
-              data.oldPrice
-            )}</p>
-            <p class="card-content__prices__price">Por: ${formatter.format(
-              data.price
-            )}</p>
-            <p class="card-content__prices__installments">ou ${
-              data.installments.count
-            }x de ${formatter.format(data.installments.value)}</p>
-            <p class="card-content__button"><a href="#" class="button button-card" value="Comprar">Comprar</a></p>
-        </div>
-    </div>
-    `;
-};
-
-const getAllProducts = (listProducts) => {
-  return listProducts.map((products) => {
-    const card = buildCard(products);
-    return (cardsGroup.innerHTML += card);
-  });
+  getAllProducts();
+  getMoreProducts(page);
 };
