+++ index.html
@@ -177,6 +177,6 @@
       </section>
     </footer>
 
-    <script src="./assets/js/scripts.js"></script>
+    <script type="module" src="./assets/js/scripts.js"></script>
   </body>
 </html>
